library_test/LibraryManager/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-simple-history>=3.10.1",
]
